# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: VIGORA - Plataforma de Gestão Jurídica
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /processos:
    get:
      operationId: listProcessos
      tags: [processos]
      summary: List processes
      parameters:
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: area
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of processes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ProcessoWithDetails"
    post:
      operationId: createProcesso
      tags: [processos]
      summary: Create process
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProcessoRequest"
      responses:
        "201":
          description: Process created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Processo"
  /processos/{id}:
    get:
      operationId: getProcesso
      tags: [processos]
      summary: Get process by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Process data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProcessoWithDetails"
        "404":
          description: Process not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateProcesso
      tags: [processos]
      summary: Update process
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProcessoRequest"
      responses:
        "200":
          description: Process updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Processo"
    delete:
      operationId: deleteProcesso
      tags: [processos]
      summary: Delete process
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Process deleted
  /prazos:
    get:
      operationId: listPrazos
      tags: [prazos]
      summary: List deadlines
      parameters:
        - name: status
          in: query
          schema:
            type: string
        - name: processoId
          in: query
          schema:
            type: integer
        - name: responsavel
          in: query
          schema:
            type: string
        - name: periodo
          in: query
          schema:
            type: string
            enum: [hoje, proximos3, proximos7, vencidos, todos]
      responses:
        "200":
          description: List of deadlines
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PrazoWithDetails"
    post:
      operationId: createPrazo
      tags: [prazos]
      summary: Create deadline
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePrazoRequest"
      responses:
        "201":
          description: Deadline created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prazo"
  /prazos/{id}:
    get:
      operationId: getPrazo
      tags: [prazos]
      summary: Get deadline by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deadline data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PrazoWithDetails"
        "404":
          description: Deadline not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updatePrazo
      tags: [prazos]
      summary: Update deadline
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePrazoRequest"
      responses:
        "200":
          description: Deadline updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Prazo"
    delete:
      operationId: deletePrazo
      tags: [prazos]
      summary: Delete deadline
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deadline deleted
...
  /notificacoes-processos/{processoId}/contatos:
    get:
      operationId: listContatosDoProcesso
      tags: [notificacoes-processos]
      summary: List WhatsApp contacts linked to a process
      parameters:
        - name: processoId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: List of linked contacts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ContatoWhatsapp"
    post:
      operationId: salvarVinculosProcesso
      tags: [notificacoes-processos]
      summary: Set (sync) WhatsApp contacts linked to a process
      parameters:
        - name: processoId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SalvarVinculosRequest"
      responses:
        "200":
          description: Links saved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VinculoSaveResult"
  /notificacoes-processos/{processoId}/contatos/{contatoId}:
    delete:
      operationId: removerContatoDoProcesso
      tags: [notificacoes-processos]
      summary: Remove a WhatsApp contact link from a process
      parameters:
        - name: processoId
          in: path
          required: true
          schema:
            type: integer
        - name: contatoId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Link removed
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    Cliente:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        cpfCnpj:
          type: string
          nullable: true
        telefone:
          type: string
          nullable: true
        email:
          type: string
          nullable: true
        observacoes:
          type: string
          nullable: true
        criadoEm:
          type: string
          format: date-time
      required:
        - id
        - nome
        - criadoEm
...
    Processo:
      type: object
      properties:
        id:
          type: integer
        numeroProcesso:
          type: string
        tribunal:
          type: string
          nullable: true
        area:
          type: string
          nullable: true
        status:
          type: string
          enum: [ativo, atencao, prazo_vencido, concluido, arquivado]
        observacoes:
          type: string
          nullable: true
        clienteId:
          type: integer
          nullable: true
        responsavelId:
          type: integer
          nullable: true
        criadoEm:
          type: string
          format: date-time
      required:
        - id
        - numeroProcesso
        - status
        - criadoEm
    ProcessoSummary:
      type: object
      properties:
        id:
          type: integer
        numeroProcesso:
          type: string
        area:
          type: string
          nullable: true
        status:
          type: string
      required:
        - id
        - numeroProcesso
        - status
    ProcessoWithDetails:
      allOf:
        - $ref: "#/components/schemas/Processo"
        - type: object
          properties:
            cliente:
              $ref: "#/components/schemas/Cliente"
              nullable: true
            proximoPrazo:
              $ref: "#/components/schemas/PrazoSummary"
              nullable: true
            prazos:
              type: array
              items:
                $ref: "#/components/schemas/Prazo"
            tarefas:
              type: array
              items:
                $ref: "#/components/schemas/Tarefa"
          required:
            - prazos
            - tarefas
    CreateProcessoRequest:
      type: object
      properties:
        numeroProcesso:
          type: string
        tribunal:
          type: string
          nullable: true
        area:
          type: string
          nullable: true
        status:
          type: string
          enum: [ativo, atencao, prazo_vencido, concluido, arquivado]
        observacoes:
          type: string
          nullable: true
        clienteId:
          type: integer
          nullable: true
        responsavelId:
          type: integer
          nullable: true
      required:
        - numeroProcesso
        - status
    Prazo:
      type: object
      properties:
        id:
          type: integer
        processoId:
          type: integer
          nullable: true
        tipo:
          type: string
          enum: [peticao, audiencia, recurso, apelacao, agravo, embargos, contestacao, cumprimento_exigencia, diligencia, outro]
        descricao:
          type: string
          nullable: true
        dataLimite:
          type: string
          format: date
        responsavel:
          type: string
          nullable: true
        prioridade:
          type: string
          enum: [alta, media, baixa]
        status:
          type: string
          enum: [ok, atencao, urgente, vencido, concluido]
        criadoEm:
          type: string
          format: date-time
      required:
        - id
        - tipo
        - dataLimite
        - prioridade
        - status
        - criadoEm
    PrazoSummary:
      type: object
      properties:
        id:
          type: integer
        tipo:
          type: string
        dataLimite:
          type: string
          format: date
        status:
          type: string
        prioridade:
          type: string
      required:
        - id
        - tipo
        - dataLimite
        - status
        - prioridade
    PrazoWithDetails:
      allOf:
        - $ref: "#/components/schemas/Prazo"
        - type: object
          properties:
            processo:
              $ref: "#/components/schemas/ProcessoSummary"
              nullable: true
    CreatePrazoRequest:
      type: object
      properties:
        processoId:
          type: integer
          nullable: true
        tipo:
          type: string
          enum: [peticao, audiencia, recurso, apelacao, agravo, embargos, contestacao, cumprimento_exigencia, diligencia, outro]
        descricao:
          type: string
          nullable: true
        dataLimite:
          type: string
          format: date
        responsavel:
          type: string
          nullable: true
        prioridade:
          type: string
          enum: [alta, media, baixa]
        status:
          type: string
          enum: [ok, atencao, urgente, vencido, concluido]
      required:
        - tipo
        - dataLimite
        - prioridade
        - status
    Tarefa:
      type: object
      properties:
        id:
          type: integer
        titulo:
          type: string
        descricao:
          type: string
          nullable: true
        processoId:
          type: integer
          nullable: true
        prazoId:
          type: integer
          nullable: true
        responsavel:
          type: string
          nullable: true
        dataLimite:
          type: string
          format: date
          nullable: true
        status:
          type: string
          enum: [pendente, em_andamento, concluida, atrasada]
        criadoEm:
          type: string
          format: date-time
      required:
        - id
        - titulo
        - status
        - criadoEm
...
    ContatoWhatsapp:
      type: object
      properties:
        id:
          type: integer
        nome:
          type: string
        numeroWhatsapp:
          type: string
        email:
          type: string
          nullable: true
        status:
          type: string
          enum: [ativo, inativo]
        observacoes:
          type: string
          nullable: true
        usuarioId:
          type: integer
          nullable: true
        criadoEm:
          type: string
          format: date-time
        atualizadoEm:
          type: string
          format: date-time
      required:
        - id
        - nome
        - numeroWhatsapp
        - status
        - criadoEm
        - atualizadoEm
...
    SalvarVinculosRequest:
      type: object
      properties:
        contatoIds:
          type: array
          items:
            type: integer
      required:
        - contatoIds
    VinculoSaveResult:
      type: object
      properties:
        success:
          type: boolean
        added:
          type: integer
        removed:
          type: integer
      required:
        - success
        - added
        - removed