# lib/api-zod/src/generated/types/createPrazoRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VIGORA - Plataforma de Gestão Jurídica
 * OpenAPI spec version: 0.1.0
 */
import type { CreatePrazoRequestPrioridade } from "./createPrazoRequestPrioridade";
import type { CreatePrazoRequestStatus } from "./createPrazoRequestStatus";
import type { CreatePrazoRequestTipo } from "./createPrazoRequestTipo";

export interface CreatePrazoRequest {
  processoId?: number | null;
  tipo: CreatePrazoRequestTipo;
  descricao?: string | null;
  dataLimite: Date;
  responsavel?: string | null;
  prioridade: CreatePrazoRequestPrioridade;
  status: CreatePrazoRequestStatus;
}