# lib/api-zod/src/generated/types/prazo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VIGORA - Plataforma de Gestão Jurídica
 * OpenAPI spec version: 0.1.0
 */
import type { PrazoPrioridade } from "./prazoPrioridade";
import type { PrazoStatus } from "./prazoStatus";
import type { PrazoTipo } from "./prazoTipo";

export interface Prazo {
  id: number;
  processoId?: number | null;
  tipo: PrazoTipo;
  descricao?: string | null;
  dataLimite: Date;
  responsavel?: string | null;
  prioridade: PrazoPrioridade;
  status: PrazoStatus;
  criadoEm: Date;
}