# lib/api-zod/src/generated/types/processo.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * VIGORA - Plataforma de Gestão Jurídica
 * OpenAPI spec version: 0.1.0
 */
import type { ProcessoStatus } from "./processoStatus";

export interface Processo {
  id: number;
  numeroProcesso: string;
  tribunal?: string | null;
  area?: string | null;
  status: ProcessoStatus;
  observacoes?: string | null;
  clienteId?: number | null;
  responsavelId?: number | null;
  criadoEm: Date;
}